required component rename and css tweaks to allow for wrapper
